# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Makana unified property and home services platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface Property {
  id: number;
  title: string;
  propertyType: string;
  listingType: string;
  price: number;
  area: number;
  rooms: number;
  bathrooms: number;
  governorate: string;
  district: string;
  finish: string;
  image: string;
  description?: string;
  featured: boolean;
  status: string;
  views?: number;
  amenities?: string[];
}

export interface PropertyInput {
  title: string;
  propertyType: string;
  listingType: string;
  price: number;
  area: number;
  rooms: number;
  bathrooms: number;
  governorate: string;
  district: string;
  finish: string;
  image: string;
  description?: string;
  amenities?: string[];
}

export interface FavoriteState {
  favorite: boolean;
}

export interface ViewingRequestInput {
  propertyId: number;
  name: string;
  phone: string;
  preferredDate: string;
  preferredTime: string;
  notes?: string;
}

export type ViewingRequest = ViewingRequestInput & {
  id: number;
  status: string;
};

export interface FinishingService {
  id: number;
  name: string;
  category: string;
  description: string;
  unit: string;
  laborPrice: number;
  accent: string;
}

export interface FinishingQuoteInput {
  serviceId: number;
  area: number;
  placeType?: string;
  bathrooms?: number;
  notes?: string;
}

export interface FinishingQuote {
  serviceId: number;
  area: number;
  laborPrice: number;
  estimatedCost: number;
  disclaimer: string;
}

export interface Professional {
  id: number;
  name: string;
  title: string;
  category: string;
  specialty: string;
  governorate: string;
  district: string;
  experience: number;
  rating: number;
  image: string;
  available: boolean;
  bio?: string;
}

export interface AppointmentInput {
  professionalId: number;
  name: string;
  phone: string;
  date: string;
  time: string;
  notes?: string;
}

export type Appointment = AppointmentInput & {
  id: number;
  status: string;
};

export interface Product {
  id: number;
  name: string;
  category: string;
  vendor: string;
  price: number;
  oldPrice: number;
  image: string;
  stock: number;
  badge: string;
  description?: string;
}

export type OrderInputItemsItem = {
  productId: number;
  quantity: number;
};

export interface OrderInput {
  name: string;
  phone: string;
  address: string;
  items: OrderInputItemsItem[];
}

export type Order = OrderInput & {
  id: number;
  status: string;
  total: number;
};

export interface DashboardSummary {
  users: number;
  properties: number;
  pendingProperties: number;
  publishedProperties?: number;
  viewingRequests: number;
  finishingRequests: number;
  professionals: number;
  products: number;
  orders: number;
  revenue?: number;
}

/**
 * Resource not found
 */
export type NotFoundResponse = Error;

export type SearchParameter = string;

export type ListPropertiesParams = {
search?: SearchParameter;
listingType?: string;
propertyType?: string;
governorate?: string;
maxPrice?: number;
featured?: boolean;
};

export type ListProfessionalsParams = {
search?: SearchParameter;
category?: string;
governorate?: string;
};

export type ListProductsParams = {
search?: SearchParameter;
category?: string;
};