# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Makana unified property and home services platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  users: number;
  properties: number;
  pendingProperties: number;
  publishedProperties?: number;
  viewingRequests: number;
  finishingRequests: number;
  professionals: number;
  products: number;
  orders: number;
  revenue?: number;
}