# lib/api-zod/src/generated/types/finishingQuoteInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Makana unified property and home services platform API
 * OpenAPI spec version: 0.1.0
 */

export interface FinishingQuoteInput {
  serviceId: number;
  area: number;
  placeType?: string;
  bathrooms?: number;
  notes?: string;
}