# lib/api-zod/src/generated/types/finishingService.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Makana unified property and home services platform API
 * OpenAPI spec version: 0.1.0
 */

export interface FinishingService {
  id: number;
  name: string;
  category: string;
  description: string;
  unit: string;
  laborPrice: number;
  accent: string;
}