# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Makana unified property and home services platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInput } from './orderInput';

export type Order = OrderInput & {
  id: number;
  status: string;
  total: number;
};