# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Makana unified property and home services platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Property {
  id: number;
  title: string;
  propertyType: string;
  listingType: string;
  price: number;
  area: number;
  rooms: number;
  bathrooms: number;
  governorate: string;
  district: string;
  finish: string;
  image: string;
  description?: string;
  featured: boolean;
  status: string;
  views?: number;
  amenities?: string[];
}