# lib/api-zod/src/generated/types/propertyInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Makana unified property and home services platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PropertyInput {
  title: string;
  propertyType: string;
  listingType: string;
  price: number;
  area: number;
  rooms: number;
  bathrooms: number;
  governorate: string;
  district: string;
  finish: string;
  image: string;
  description?: string;
  amenities?: string[];
}