# lib/api-zod/src/generated/types/viewingRequestInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Makana unified property and home services platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ViewingRequestInput {
  propertyId: number;
  name: string;
  phone: string;
  preferredDate: string;
  preferredTime: string;
  notes?: string;
}